5. Generalize component to use without stars
