+++ src/App.js
@@ -1,5 +1,5 @@
 import "bootstrap/dist/css/bootstrap.css";
-import StarRating from "./StarRating";
+import Rating from "./Rating";
 import { useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faStar } from "@fortawesome/free-solid-svg-icons";
@@ -8,7 +8,7 @@ export default function App() {
   const [rating, setRating] = useState(3);
 
   return (
-    <StarRating
+    <Rating
       rating={rating}
       onClick={(newRating) => {
         setRating(newRating);
@@ -25,6 +25,6 @@ export default function App() {
           </button>
         );
       }}
-    </StarRating>
+    </Rating>
   );
 }
